Split Symbol.swift to reduce complexity of the file
