+++ Sources/SymbolKit/SymbolGraph/LineList/SourceRange/SourceRange+Position.swift
@@ -1,0 +1,28 @@
+/*
+ This source file is part of the Swift.org open source project
+
+ Copyright (c) 2021 Apple Inc. and the Swift project authors
+ Licensed under Apache License v2.0 with Runtime Library Exception
+
+ See https://swift.org/LICENSE.txt for license information
+ See https://swift.org/CONTRIBUTORS.txt for Swift project authors
+*/
+
+import Foundation
+
+extension SymbolGraph.LineList.SourceRange {
+    /**
+     Represents a cursor position in text.
+     */
+    public struct Position: Equatable, Codable {
+        /**
+         The zero-based line number of a document.
+         */
+        public var line: Int
+
+        /**
+         The zero-based *byte offset* into a line.
+         */
+        public var character: Int
+    }
+}
