+++ Sources/SymbolKit/SymbolGraph/Symbol/Symbol+AccessControl.swift
@@ -1,0 +1,22 @@
+/*
+ This source file is part of the Swift.org open source project
+
+ Copyright (c) 2021 Apple Inc. and the Swift project authors
+ Licensed under Apache License v2.0 with Runtime Library Exception
+
+ See https://swift.org/LICENSE.txt for license information
+ See https://swift.org/CONTRIBUTORS.txt for Swift project authors
+*/
+
+import Foundation
+
+extension SymbolGraph.Symbol {
+    /// The language-agnostic access control of a symbol,
+    /// such as "public", "private", "protected".
+    public struct AccessControl: Codable, RawRepresentable, Hashable {
+        public var rawValue: String
+        public init(rawValue: String) {
+            self.rawValue = rawValue
+        }
+    }
+}
