+++ Sources/SymbolKit/UnifiedSymbolGraph/GraphCollector.swift
